refactor(input): split backends and resolver into focused modules (#707 a)
Behavior-preserving refactor — no public tool changes, no fallback
semantics altered.

New modules under src/input/:
  - backend.ts          — InputBackend interface + InputBackendKind type
  - simctl-backend.ts   — SimctlInputBackend
  - applescript-backend.ts — AppleScriptInputBackend + key maps (opt-in guard preserved)
  - webkit-backend.ts   — WebKitInputBackend + key maps
  - flutter-resolver.ts — FlutterVMResolverInstance (per-instance cache)
  - backend-resolver.ts — InputBackendResolver class (instance-owned state,
                          no module globals) + HeadlessInputUnavailableError

src/tools/native-input-backend.ts is now a compatibility shim: every
previously-public symbol is re-exported so existing callers (native-input-utils,
cache-budget, app-alert-handle, app-scroll-native, app-dismiss-keyboard)
continue to work without modification.

sim-hid-input-backend.ts, flutter-vm-input-backend.ts, and
pointer-service-input-backend.ts now import InputBackend from
src/input/backend (no cycle through the shim). Metrics modules likewise
import InputBackendKind directly from src/input/backend.

Note on sim-hid-input-backend.ts placement: left in src/tools/ to
minimize diff scope per the issue guidance ("your choice: move it under
src/input/ for consistency OR leave in place and adjust imports"). The
import was updated to src/input/backend — no behavioral change.

Tests: 80 existing native-input-backend tests pass unchanged.
New: tests/unit/input-backend-resolver.test.ts — 14 tests covering
instance isolation, reset() cache clearing, and fallback order assertions.
